Add files via upload
greater lag on sql query to see if it fixes the seldom database hang up issues when cpu is taxed, it backs up then errors out. as more of these errors accumulate it causes the database to lock out and need to be restarted to function properly again for writing to the database.